Move apps-manifest outside of edit folder
