+++ src/app/apps-manifest.ts
@@ -1,32 +1,32 @@
 import { StaticImageData } from 'next/image';
 
-import amazon from '../../../public/app-icons/amazon.png';
-import asana from '../../../public/app-icons/asana.png';
-import bamboohr from '../../../public/app-icons/bamboohr.png';
-import bestbuy from '../../../public/app-icons/bestbuy.png';
-import browserstack from '../../../public/app-icons/browserstack.png';
-import googlecalendar from '../../../public/app-icons/googlecalendar.png';
-import confluence from '../../../public/app-icons/confluence.png';
-import craigslist from '../../../public/app-icons/craigslist.png';
-import googledocs from '../../../public/app-icons/googledocs.png';
-import googledrive from '../../../public/app-icons/googledrive.png';
-import ebay from '../../../public/app-icons/ebay.png';
-import facebook from '../../../public/app-icons/facebook.png';
-import gmail from '../../../public/app-icons/gmail.png';
-import google from '../../../public/app-icons/google.png';
-import jira from '../../../public/app-icons/jira.png';
-import memoryexpress from '../../../public/app-icons/memoryexpress.png';
-import messenger from '../../../public/app-icons/messenger.png';
-import netflix from '../../../public/app-icons/netflix.png';
-import googlephotos from '../../../public/app-icons/googlephotos.png';
-import slack from '../../../public/app-icons/slack.png';
-import soundcloud from '../../../public/app-icons/soundcloud.png';
-import googlesheets from '../../../public/app-icons/googlesheets.png';
-import used from '../../../public/app-icons/used.png';
-import worldtimebuddy from '../../../public/app-icons/worldtimebuddy.png';
-import youtube from '../../../public/app-icons/youtube.png';
-import zendesk from '../../../public/app-icons/zendesk.png';
-import zoom from '../../../public/app-icons/zoom.png';
+import amazon from '../../public/app-icons/amazon.png';
+import asana from '../../public/app-icons/asana.png';
+import bamboohr from '../../public/app-icons/bamboohr.png';
+import bestbuy from '../../public/app-icons/bestbuy.png';
+import browserstack from '../../public/app-icons/browserstack.png';
+import googlecalendar from '../../public/app-icons/googlecalendar.png';
+import confluence from '../../public/app-icons/confluence.png';
+import craigslist from '../../public/app-icons/craigslist.png';
+import googledocs from '../../public/app-icons/googledocs.png';
+import googledrive from '../../public/app-icons/googledrive.png';
+import ebay from '../../public/app-icons/ebay.png';
+import facebook from '../../public/app-icons/facebook.png';
+import gmail from '../../public/app-icons/gmail.png';
+import google from '../../public/app-icons/google.png';
+import jira from '../../public/app-icons/jira.png';
+import memoryexpress from '../../public/app-icons/memoryexpress.png';
+import messenger from '../../public/app-icons/messenger.png';
+import netflix from '../../public/app-icons/netflix.png';
+import googlephotos from '../../public/app-icons/googlephotos.png';
+import slack from '../../public/app-icons/slack.png';
+import soundcloud from '../../public/app-icons/soundcloud.png';
+import googlesheets from '../../public/app-icons/googlesheets.png';
+import used from '../../public/app-icons/used.png';
+import worldtimebuddy from '../../public/app-icons/worldtimebuddy.png';
+import youtube from '../../public/app-icons/youtube.png';
+import zendesk from '../../public/app-icons/zendesk.png';
+import zoom from '../../public/app-icons/zoom.png';
 
 export type AppTypes = {
   name: string;
